Add safe script to fix tmux in macOS
This script is used to fix tmux in macOS. It is used to fix the issue
where tmux does not have access to the system clipboard. This script
checks if the `reattach-to-user-namespace` command is available and
runs it if it is. If it is not available, it just runs the command.

https://github.com/jimeh/dotfiles/commit/3838db8